Refactor: extract the other half of setup function
